QuantEngine/requirements.txt
# Core data science and trading libraries
pandas>=1.5.0
numpy>=1.21.0
polars>=0.15.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Machine learning and optimization
xgboost>=1.6.0
lightgbm>=3.3.0
statsmodels>=0.13.0
optuna>=2.10.0

# Backtesting and technical analysis
backtrader>=1.9.0
ta-lib>=0.4.0
quantlib-python>=1.28

# Data storage and processing
pyarrow>=10.0.0
duckdb>=0.6.0
sqlalchemy>=1.4.0

# API and web services
requests>=2.28.0
aiohttp>=3.8.0
yfinance>=0.1.70
alpha-vantage>=2.3.0

# Configuration and validation
pydantic>=1.9.0
hydra-core>=1.2.0
omegaconf>=2.2.0

# Experiment tracking and MLflow
mlflow>=1.26.0
wandb>=0.12.0

# Visualization and reporting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.8.0
jinja2>=3.0.0
reportlab>=3.6.0

# Performance and parallel processing
numba>=0.56.0
joblib>=1.1.0
dask>=2022.6.0

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
mypy>=0.950

# LLM and AI (optional - can use local models)
openai>=0.27.0
transformers>=4.20.0
sentence-transformers>=2.2.0

# Logging and utilities
loguru>=0.6.0
tqdm>=4.64.0
python-dotenv>=0.19.0
